Rename : Edit assets dir & path
