+++ src/layouts/Header.tsx
@@ -1,5 +1,5 @@
 import { Link, useLocation } from 'react-router-dom';
-import logo from '../assets/logo.png';
+import logo from '../assets/images/logo.png';
 import { AiOutlineUser } from 'react-icons/ai';
 
 const Header = () => {
